+++ src/components/mypage/admin/ReviewList/ReviewListSekelton.tsx
@@ -1,5 +1,25 @@
 import Skeleton from "react-loading-skeleton";
+import itemStyles from "../ReviewListItem/ReviewListItem.module.css";
 import styles from "./ReviewList.module.css";
+
+const ReviewListItemSkeleton = () => {
+  return (
+    <>
+      <div className={itemStyles["review-contents"]}>
+        <Skeleton height={"0.7rem"} width={120} />
+        <div style={{ marginTop: "0.625rem" }}>
+          <Skeleton height={"0.875rem"} width={"100%"} />
+        </div>
+        <div style={{ marginLeft: "auto" }}>
+          <div>
+            <Skeleton height={"0.75rem"} width={100} />
+          </div>
+        </div>
+      </div>
+      <div style={{ width: "2rem", height: "2rem" }}></div>
+    </>
+  );
+};
 
 const ReviewListSekelton = () => {
   return (
@@ -9,18 +29,7 @@ const ReviewListSekelton = () => {
       </div>
       <ul className={styles["review-list"]}>
         <div className={styles["review"]}>
-          <div className={styles["review-contents"]}>
-            <Skeleton height={"0.7rem"} width={120} />
-            <div style={{ marginTop: "0.625rem" }}>
-              <Skeleton height={"0.875rem"} width={"100%"} />
-            </div>
-            <div style={{ marginLeft: "auto" }} className={styles["review__date"]}>
-              <div>
-                <Skeleton height={"0.75rem"} width={100} />
-              </div>
-            </div>
-          </div>
-          <div style={{ width: "2rem", height: "2rem" }}></div>
+          <ReviewListItemSkeleton />
         </div>
       </ul>
     </>
