+++ src/components/mypage/index.tsx
@@ -11,6 +11,7 @@ import StoryItem from "./user/StoryItem/StoryItem";
 import ShowList from "./admin/ShowList/ShowList";
 import ShowListItem from "./admin/ShowListItem/ShowListItem";
 import ReviewList from "./admin/ReviewList/ReviewList";
+import ReviewListItem from "./admin/ReviewListItem/ReviewListItem";
 import LikeItemList from "./user/ListItemList/LikeItemList";
 import LikeItem from "./user/LikeItem/LikeItem";
 
@@ -28,6 +29,7 @@ export {
   ShowList,
   ShowListItem,
   ReviewList,
+  ReviewListItem,
   LikeItemList,
   LikeItem,
 };
